Fixed missing method due to deprecation handling and stuff
