+++ app/src/androidTest/java/com/itachi1706/cheesecakeutilities/BinHexModuleActivityTest.kt
@@ -31,13 +31,7 @@ class BinHexModuleActivityTest {
 
     @Before
     fun unlockScreen() {
-        val activity = mActivityTestRule.activity
-        val wakeUpDevice = {
-            activity.setTurnScreenOn(true)
-            activity.setShowWhenLocked(true)
-            activity.window.addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON)
-        }
-        activity.runOnUiThread(wakeUpDevice)
+        TestHelper.wakeUpDevice(mActivityTestRule)
     }
 
     @Test
