+++ app/src/androidTest/java/com/itachi1706/cheesecakeutilities/HtcSerialIdentificationTest.kt
@@ -42,13 +42,7 @@ class HtcSerialIdentificationTest {
 
     @Before
     fun unlockScreen() {
-        val activity = mActivityTestRule.activity
-        val wakeUpDevice = {
-            activity.setTurnScreenOn(true)
-            activity.setShowWhenLocked(true)
-            activity.window.addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON)
-        }
-        activity.runOnUiThread(wakeUpDevice)
+        TestHelper.wakeUpDevice(mActivityTestRule)
     }
 
     @Test
